+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+- 2024-10-16: added missing span stats to SQL repository implementation.
+
 ## [4.1.1] - 2024-10-16
 
 - 2024-10-16: